Project_Starter_Files-Building_a_Metrics_Dashboard/reference-app/backend/requirements.txt
pymongo==3.11.0
Flask-PyMongo==2.3.0
Flask==1.1.2
requests==2.23.0
pylint==2.4.0
gunicorn==20.0.4
prometheus-flask-exporter==0.18.2
opentelemetry-exporter-jaeger==1.5.0
opentelemetry-instrumentation-flask==0.24b0
jaeger-client==4.3.0
Flask-OpenTracing==1.1.0
flask-cors==3.0.10